test(web): consolidate duplicated master-bus test stubs into web/test/
Pure extraction, no behavior/assertion changes: the fake Web Audio node
classes (biquad/compressor/splitter/analyser) were byte-identical across
hooks/use-studio-playback.test.tsx and app/studio/page.test.tsx; stubRaf()
was byte-identical (modulo one comment) across those two plus
MasterBusPanel.test.tsx and StereoMeter.test.tsx. Both now live in
web/test/, matching the existing clip-factory.ts/router-mock.ts convention.
